correção de bug - tipo da quantidade de ingrediente

+++ app/src/main/java/com/empsoft/safe_meal/services/retrofit_models/ExtendedIngredient.java
@@ -20,7 +20,7 @@ public class ExtendedIngredient {
     private String name;
 
     @SerializedName("amount")
-    private Integer amount;
+    private float amount;
 
     @SerializedName("unit")
     private String unit;
@@ -41,7 +41,7 @@ public class ExtendedIngredient {
 
     public ExtendedIngredient() {}
 
-    public ExtendedIngredient(Integer id, String aisle, String image, String name, Integer amount, String unit, String unitShort, String unitLong, String originalString, List<String> metaInfortmation) {
+    public ExtendedIngredient(Integer id, String aisle, String image, String name, float amount, String unit, String unitShort, String unitLong, String originalString, List<String> metaInfortmation) {
         this.id = id;
         this.aisle = aisle;
         this.image = image;
@@ -70,7 +70,7 @@ public class ExtendedIngredient {
         return name;
     }
 
-    public Integer getAmount() {
+    public float getAmount() {
         return amount;
     }
